Trying to sort out notion API call
