+++ package.json
@@ -10,6 +10,7 @@
     "format": "prettier --write ."
   },
   "dependencies": {
+    "@notionhq/client": "^2.2.15",
     "@radix-ui/react-dialog": "^1.0.4",
     "@radix-ui/react-tooltip": "^1.0.6",
     "@types/lodash": "^4.14.197",
@@ -25,11 +26,9 @@
     "lodash": "^4.17.21",
     "lodash.merge": "^4.6.2",
     "next": "^14.2.6",
-    "notion-client": "^6.16.0",
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "react-icons": "^4.11.0",
-    "react-notion-x": "^6.16.0",
     "sharp": "^0.32.5",
     "typescript": "5.1.6"
   },
