+++ src/components/SocialLinks/WithoutTooltip.tsx
@@ -1,24 +1,8 @@
-import { isValidElement } from "react";
 import Link from "next/link";
 import { VisuallyHidden } from "@/components";
 import styles from "./socialLinks.styles.module.css";
 import type { SocialLinkProps } from "./socialLinks.types";
-import { IconType } from "react-icons";
-
-interface IconProps {
-  icon: IconType | JSX.Element;
-  size: number;
-}
-
-const Icon = ({ icon, size }: IconProps) => {
-  if (typeof icon === "function") {
-    const Icon = icon;
-    return <Icon aria-hidden="true" style={{ fontSize: `${size}px` }} />;
-  } else if (isValidElement(icon)) {
-    return icon;
-  }
-  return null;
-};
+import { TooltipIcon } from "./TooltipIcon";
 
 export const WithoutTooltip = ({ link, size }: SocialLinkProps) => {
   return (
@@ -29,7 +13,7 @@ export const WithoutTooltip = ({ link, size }: SocialLinkProps) => {
       {...(link.attributes ? link.attributes : {})}
     >
       <>
-        <Icon icon={link.icon} size={size} />
+        <TooltipIcon icon={link.icon} size={size} />
         <VisuallyHidden>{link.screenReader}</VisuallyHidden>
       </>
     </Link>
